requirements.txt
Django~=5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0
psycopg2-binary==2.9.9
pytest~=8.3.3
pytest-django==4.9.0
drf-yasg==1.21.8